--- src/components/atoms/signUp/primaryHeader/PrimaryTitle.tsx
@@ -1,7 +1,0 @@
-import { PrimaryTitleSpan } from "./PrimaryTitleStyles";
-
-const PrimaryTitle = () => {
-  return <PrimaryTitleSpan>제주디어스 회원가입</PrimaryTitleSpan>;
-};
-
-export default PrimaryTitle;
