+++ src/components/atoms/signUp/primaryHeader/PrimaryTitleKorean.tsx
@@ -1,0 +1,7 @@
+import { PrimaryTitleKoreanSpan } from "./PrimaryTitleKoreanStyles";
+
+const PrimaryTitleKorean = () => {
+  return <PrimaryTitleKoreanSpan>제주디어스 회원가입</PrimaryTitleKoreanSpan>;
+};
+
+export default PrimaryTitleKorean;
