+++ src/components/atoms/signUp/primaryHeader/PrimaryTitleKoreanStyles.ts
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import tw from "twin.macro";
 
-export const PrimaryTitleSpan = styled.span`
+export const PrimaryTitleKoreanSpan = styled.span`
   ${tw`
     text-5xl
     text-[#FFFFFF]
