+++ src/components/molecules/signUp/primaryHeader/PrimaryTitleHeader.tsx
@@ -1,0 +1,14 @@
+import PrimaryTitleEnglish from "~/components/atoms/signUp/primaryHeader/PrimaryTitleEnglish";
+import { PrimaryTitleHeaderBox } from "./PrimaryTitleHeaderStyles";
+import PrimaryTitleKorean from "~/components/atoms/signUp/primaryHeader/PrimaryTitleKorean";
+
+const PrimaryTitleHeader = () => {
+  return (
+    <PrimaryTitleHeaderBox>
+      <PrimaryTitleEnglish />
+      <PrimaryTitleKorean />
+    </PrimaryTitleHeaderBox>
+  );
+};
+
+export default PrimaryTitleHeader;
