+++ src/components/organisms/signUp/primaryHeader/PrimaryHeader.tsx
@@ -1,0 +1,12 @@
+import PrimaryTitleHeader from "~/components/molecules/signUp/primaryHeader/PrimaryTitleHeader";
+import { PrimaryHeaderBox } from "./PrimaryHeaderStyles";
+
+const PrimaryHeader = () => {
+  return (
+    <PrimaryHeaderBox>
+      <PrimaryTitleHeader />
+    </PrimaryHeaderBox>
+  );
+};
+
+export default PrimaryHeader;
