--- src/components/organisms/signUp/primaryHeader/PrimaryHeaderContent.tsx
@@ -1,12 +1,0 @@
-import PrimaryHeader from "~/components/molecules/signUp/primaryHeader/PrimaryHeader";
-import { PrimaryHeaderContentBox } from "./PrimaryHeaderContentStyles";
-
-const PrimaryHeaderContent = () => {
-  return (
-    <PrimaryHeaderContentBox>
-      <PrimaryHeader />
-    </PrimaryHeaderContentBox>
-  );
-};
-
-export default PrimaryHeaderContent;
